% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcentrality.R
\name{ggcentrality}
\alias{ggcentrality}
\title{Calculate centrality indices for vertices linked to a sf object}
\usage{
ggcentrality(
  n,
  m,
  lkp = NULL,
  m_name = NULL,
  n_name = "vertex.names",
  par.deg = NULL,
  par.bet = NULL,
  par.clo = NULL
)
}
\arguments{
\item{n}{A \code{network} or \code{igraph} object.}

\item{m}{A \code{sf} object.}

\item{lkp}{An optional lookup table.}

\item{m_name}{Optional \code{character}, name of field in \code{m} and of
column in \code{lkp}.}

\item{n_name}{Optional \code{character}, name of vertex attribute in \code{n}
and of column in \code{lkp}.}

\item{par.deg}{List with additional optional parameters to functions
\code{\link[sna]{degree}} or \code{\link[igraph]{degree}}.}

\item{par.bet}{List with additional optional parameters to functions
\code{\link[sna]{betweenness}} or \code{\link[igraph]{betweenness}}.}

\item{par.clo}{List with additional optional parameters to functions
\code{\link[sna]{closeness}} or \code{\link[igraph]{closeness}}.}
}
\value{
An \code{sf} object, input \code{m} with added columns for centrality
indices (degree, betweenness, closeness; existing columns with the same name
will be overwritten) and with only the features linked to vertices in input
\code{n}.
}
\description{
Given a \code{sf} object with features that can be linked to a \code{network}
or \code{igraph} object, obtain centrality indices for linked features.
}
