% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_network_map.R
\name{is_lookup_table}
\alias{is_lookup_table}
\title{Is data frame a lookup table?}
\usage{
is_lookup_table(lkp, m_name = NULL, n_name = NULL)
}
\arguments{
\item{lkp}{A \code{data.frame}.}

\item{m_name}{Optional, a \code{character} string with the name of the column
in \code{lkp} to check against \code{m}.}

\item{n_name}{Optional, a \code{character} string with the name of the column
in \code{lkp} to check against \code{n}.}
}
\value{
\code{FALSE} on error, a vector with \code{m_name} and \code{n_name}
if the lookup table is valid.
}
\description{
Checks whether a \code{data.frame} is a valid lookup table.
}
\examples{
# lkptbl=data.frame(id=c("a1", "b2", "c3", "d4"), name=c("a", "b", "c", "d"))
# netmap::is_lookup_table(lkptbl) #valid
# lkptbl2=data.frame(id=c("a1", "b2", "c3", "d4"), name=c("a", "b", NA, "d"))
# netmap::is_lookup_table(lkptbl2) #invalid, NA
# lkptbl3=data.frame(id=c("a1", "b2", "c3", "d4"), name=c("a", "b", "b", "d"))
# netmap::is_lookup_table(lkptbl3) #invalid, duplicates
}
