% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_network_map.R
\name{is_network}
\alias{is_network}
\title{Is object a network?}
\usage{
is_network(n)
}
\arguments{
\item{n}{Object of class \code{network} or \code{igraph}.}
}
\value{
\code{TRUE} if object of class \code{network}, \code{FALSE} otherwise.
}
\description{
Checks whether an object is a \code{network} object or an \code{igraph} object,
returns message if it's not
}
\examples{
# is_network(network::network(1)) #TRUE
}
