% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_network_map.R
\name{is_sf}
\alias{is_sf}
\title{Is object a map?}
\usage{
is_sf(m)
}
\arguments{
\item{m}{Object of class \code{sf}.}
}
\value{
\code{TRUE} if object of classes \code{sf} and \code{data.frame}, \code{FALSE} otherwise.
}
\description{
Checks whether an object is an \code{sf} object, returns message if it's not
}
