% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_network_map.R
\name{link_network_map}
\alias{link_network_map}
\title{Link a network and a map}
\usage{
link_network_map(m, n, m_name, n_name = "vertex.names")
}
\arguments{
\item{m}{Object of class \code{sf}.}

\item{n}{Object of class \code{network} or \code{igraph}.}

\item{m_name}{Name of the map field to use for the link.}

\item{n_name}{Name of the vertex attribute to use for the link, defaults to
\code{vertex.names}.}
}
\value{
On success a list with two vectors, one of features in \code{m}
present in \code{n}, the other of nodes in \code{n} present in \code{m},
\code{-1} on error.
}
\description{
Checks which vertices of a \code{network} object can be represented with
features of a \code{sf} object.
}
