% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmap-package.R
\docType{package}
\name{netmap-package}
\alias{netmap-package}
\title{netmap: Represent Network Objects on a Map}
\description{
Represent 'network' or 'igraph' objects whose vertices can be represented by features in an 'sf' object as a network graph surmising a 'sf' plot. Fits into 'ggplot2' grammar.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/artod83/netmap}
  \item Report bugs at \url{https://github.com/artod83/netmap/issues}
}

}
\author{
\strong{Maintainer}: Matteo Dimai \email{matteo.dimai@phd.units.it} (\href{https://orcid.org/0000-0003-1126-5234}{ORCID})

}
\keyword{internal}
