% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_map.R
\name{reduce_to_map}
\alias{reduce_to_map}
\title{Reduces network to vertices present on the map}
\usage{
reduce_to_map(n, link, n_name)
}
\arguments{
\item{n}{A \code{network} or \code{igraph} object.}

\item{link}{A vector with the identifiers of the vertices to keep.}

\item{n_name}{Name of the vertex attribute to filter on.}
}
\value{
A \code{network} or \code{igraph} object with only the vertices
listed in \code{link}.
}
\description{
Removes vertices from a \code{network} or \code{igraph} object which are not
present in the link vector produced by \code{\link{link_network_map}} or
\code{\link{link_network_map2}}.
}
\examples{
# net=network(matrix(c(0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0), nrow=4, byrow=TRUE))
# network::set.vertex.attribute(net, "name", value=c("a", "b", "c", "d"))
# wkb = structure(list("01010000204071000000000000801A064100000000AC5C1641",
# "01010000204071000000000000801A084100000000AC5C1441",
# "01010000204071000000000000801A044100000000AC5C1241",
# "01010000204071000000000000801A024100000000AC5C1841"), class = "WKB")
# map=st_sf(id=c("a", "b", "c", "e"), st_as_sfc(wkb, EWKB=TRUE))
# link=link_network_map(map, net, "id", "name")
# net2=reduce_to_map(net, link, "name") #a net with vertices a, b and c
}
