% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_network_map.R
\name{is_lookup_table}
\alias{is_lookup_table}
\title{Is data frame a lookup table?}
\usage{
is_lookup_table(lkp, m_name = NULL, n_name = NULL)
}
\arguments{
\item{lkp}{A \code{data.frame}.}

\item{m_name}{Optional, a \code{character} string with the name of the column
in \code{lkp} to check against \code{m}.}

\item{n_name}{Optional, a \code{character} string with the name of the column
in \code{lkp} to check against \code{n}.}
}
\value{
\code{FALSE} on error, a vector with \code{m_name} and \code{n_name}
if the lookup table is valid.
}
\description{
Checks whether a \code{data.frame} is a valid lookup table.
}
