% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_network_map.R
\name{is_sf}
\alias{is_sf}
\title{Is object a map?}
\usage{
is_sf(m)
}
\arguments{
\item{m}{Object of class \code{sf}.}
}
\value{
\code{TRUE} if object of classes \code{sf} and \code{data.frame}, \code{FALSE} otherwise.
}
\description{
Checks whether an object is an \code{sf} object, returns message if it's not
}
\examples{
# wkb = structure(list("01010000204071000000000000801A064100000000AC5C1641",
# "01010000204071000000000000801A084100000000AC5C1441",
# "01010000204071000000000000801A044100000000AC5C1241",
# "01010000204071000000000000801A024100000000AC5C1841"), class = "WKB")
# map=st_sf(id=c("a1", "b2", "c3", "d4"), st_as_sfc(wkb, EWKB=TRUE))
# is_sf(map) #TRUE
}
