% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_network_map.R
\name{link_network_map}
\alias{link_network_map}
\title{Link a network and a map}
\usage{
link_network_map(m, n, m_name, n_name = "vertex.names")
}
\arguments{
\item{m}{Object of class \code{sf}.}

\item{n}{Object of class \code{network}.}

\item{m_name}{Name of the map field to use for the link.}

\item{n_name}{Name of the vertex attribute to use for the link, defaults to
\code{vertex.names}.}
}
\value{
On success a list with two vectors, one of features in \code{m}
present in \code{n}, the other of nodes in \code{n} present in \code{m},
\code{-1} on error.
}
\description{
Checks which vertices of a \code{network} object can be represented with
features of a \code{sf} object.
}
\examples{
# net=network(matrix(c(0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0), nrow=4, byrow=TRUE))
# network::set.vertex.attribute(net, "name", value=c("a", "b", "c", "d"))
# wkb = structure(list("01010000204071000000000000801A064100000000AC5C1641",
# "01010000204071000000000000801A084100000000AC5C1441",
# "01010000204071000000000000801A044100000000AC5C1241",
# "01010000204071000000000000801A024100000000AC5C1841"), class = "WKB")
# map=st_sf(id=c("a", "b", "c", "d"), st_as_sfc(wkb, EWKB=TRUE))
# link_network_map(map, net, "id", "name")
}
